requirements.txt
fastapi
uvicorn[standard]
pydantic
chromadb
sentence-transformers
openai
anthropic
torch
python-dotenv
